Merge pull request #833 from moninom1/newdev

+++ source/FreeRTOS_IPv4.c
@@ -58,8 +58,8 @@
 /**
  * @brief Check IPv4 packet length.
  *
- * @param[in] pvEthernetBuffer: The Ethernet packet received.
- * @param[in] uxBufferLength: The total number of bytes received.
+ * @param[in] pvEthernetBuffer The Ethernet packet received.
+ * @param[in] uxBufferLength The total number of bytes received.
  *
  * @return pdPASS when the length fields in the packet OK, pdFAIL when the packet
  *         should be dropped.
@@ -101,7 +101,7 @@
                 break;
             }
 
-            ucVersionHeaderLength = ( ucVersionHeaderLength & ( uint8_t ) 0x0FU ) << 2;
+            ucVersionHeaderLength = ( uint16_t ) ( ( ucVersionHeaderLength & ( uint8_t ) 0x0FU ) << 2U );
             uxIPHeaderLength = ( UBaseType_t ) ucVersionHeaderLength;
 
             /* Check if the complete IP-header is transferred. */
@@ -183,7 +183,7 @@
 /**
  * @brief Is the IP address an IPv4 multicast address.
  *
- * @param[in] ulIPAddress: The IP address being checked.
+ * @param[in] ulIPAddress The IP address being checked.
  *
  * @return pdTRUE if the IP address is a multicast address or else, pdFALSE.
  */
@@ -208,9 +208,9 @@ BaseType_t xIsIPv4Multicast( uint32_t ulIPAddress )
 /**
  * @brief Check whether this IPv4 packet is to be allowed or to be dropped.
  *
- * @param[in] pxIPPacket: The IP packet under consideration.
- * @param[in] pxNetworkBuffer: The whole network buffer.
- * @param[in] uxHeaderLength: The length of the header.
+ * @param[in] pxIPPacket The IP packet under consideration.
+ * @param[in] pxNetworkBuffer The whole network buffer.
+ * @param[in] uxHeaderLength The length of the header.
  *
  * @return Whether the packet should be processed or dropped.
  */
@@ -408,7 +408,7 @@ eFrameProcessingResult_t prvAllowIPPacketIPv4( const IPPacket_t * const pxIPPack
 
 
 /** @brief Check if the IP-header is carrying options.
- * @param[in] pxNetworkBuffer: the network buffer that contains the packet.
+ * @param[in] pxNetworkBuffer the network buffer that contains the packet.
  *
  * @return Either 'eProcessBuffer' or 'eReleaseBuffer'
  */
